+++ tango/integrations/pytorch_lightning/callbacks.py
@@ -3,7 +3,7 @@ import pytorch_lightning as pl
 from tango.common.registrable import Registrable
 
 
-class LightningCallback(pl.callbacks.base.Callback, Registrable):
+class LightningCallback(pl.Callback, Registrable):
     """
     This is simply a :class:`~tango.common.registrable.Registrable`
     version of the PyTorch Lightning :class:`~pytorch_lightning.callbacks.base.Callback`.
@@ -12,9 +12,5 @@ class LightningCallback(pl.callbacks.base.Callback, Registrable):
 
 # Register all callbacks.
 for name, cls in pl.callbacks.__dict__.items():
-    if (
-        isinstance(cls, type)
-        and issubclass(cls, pl.callbacks.base.Callback)
-        and not cls == pl.callbacks.base.Callback
-    ):
+    if isinstance(cls, type) and issubclass(cls, pl.Callback) and not cls == pl.Callback:
         LightningCallback.register("pytorch_lightning::" + name)(cls)