+++ tango/integrations/transformers/__init__.py
@@ -66,6 +66,8 @@ library under the corresponding class from the `torch <torch.html>`_ integration
             transformers::AutoModelForImageClassification::from_pretrained
             transformers::AutoModelForImageSegmentation::from_config
             transformers::AutoModelForImageSegmentation::from_pretrained
+            transformers::AutoModelForInstanceSegmentation::from_config
+            transformers::AutoModelForInstanceSegmentation::from_pretrained
             transformers::AutoModelForMaskedImageModeling::from_config
             transformers::AutoModelForMaskedImageModeling::from_pretrained
             transformers::AutoModelForMaskedLM::from_config