+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added an `ExecutorOutput` dataclass that is returned by `Executor.execute_step_graph()`.
 - `StepGraph` now prints itself in a readable way.
 - Tango now automatically detects when it's running under a debugger, and disables multicore support accordingly. Many debuggers can't properly follow sub-processes, so this is a convenience for people who love debuggers.
+- Added more models to the stuff we can import from the transformers library.
 - Added new example for finetuning text-to-text models.
 
 ### Changed
@@ -33,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where `TorchTrainStep` working directory would be left in an unrecoverable state if training failed after saving the final model weights.
 - Fixed bug in `FromParams` where `**kwargs` might be passed down to the constructors of arguments.
 - Fixed bug in the way dependencies are tracked between steps
+- Compatibility with PyTorch Lightning 1.6
 
 
 ## [v0.6.0](https://github.com/allenai/tango/releases/tag/v0.6.0) - 2022-02-25